+++ apps/web/src/messages/talentRequestMessages.ts
@@ -128,6 +128,17 @@ const messages = defineMessages({
     description:
       "Header for the number of user skills matching requested skills",
   },
+  changeStatus: {
+    defaultMessage: "Change status to {status}",
+    id: "DVWX9O",
+    description:
+      "Label for the change status action in the talent request tracked users table",
+  },
+  referred: {
+    defaultMessage: "Referred",
+    id: "ZAHz5C",
+    description: "Status for referred candidates",
+  },
 });
 
 export default messages;
