+++ apps/web/src/pages/TalentRequests/components/TalentRequestTrackedUsersInbox/dialogs/NotSelectTrackedUsersDialog.tsx
@@ -19,7 +19,6 @@ import { toast } from "@gc-digital-talent/toast";
 import { unpackMaybes } from "@gc-digital-talent/helpers";
 
 import messages from "../messages";
-
 import StatusDialogLayout from "./StatusDialogLayout";
 import { type StatusDialogProps, trackedUsersMutationContext } from "./shared";
