+++ apps/web/src/pages/TalentRequests/components/TalentRequestTrackedUsersInbox/dialogs/ReferTrackedUsersDialog.tsx
@@ -8,7 +8,6 @@ import { formMessages } from "@gc-digital-talent/i18n";
 import { toast } from "@gc-digital-talent/toast";
 
 import messages from "../messages";
-
 import StatusDialogLayout from "./StatusDialogLayout";
 import { type StatusDialogProps, trackedUsersMutationContext } from "./shared";
