--- apps/web/src/pages/TalentRequests/components/TalentRequestTrackedUsersTable/messages.ts
@@ -1,15 +1,0 @@
-import { defineMessages } from "react-intl";
-
-export default defineMessages({
-  changeStatus: {
-    defaultMessage: "Change status to {status}",
-    id: "DVWX9O",
-    description:
-      "Label for the change status action in the talent request tracked users table",
-  },
-  referred: {
-    defaultMessage: "Referred",
-    id: "ZAHz5C",
-    description: "Status for referred candidates",
-  },
-});
